Fixed the folder structure according to industry standards
