+++ frontend/src/App.jsx
@@ -5,22 +5,22 @@ import Dessert from "./pages/dessert";
 import Nonveg from "./pages/nonveg";
 import Beverages from "./pages/beverages";
 import "./styles/animations.css";
-import PaneerButterMasala from "./pages/PaneerButterMasala";
-import DalMaKhani from "./pages/DalMakhani";
-import PalakSoup from "./pages/PalakSoup";
-import VegetableBiryani from "./pages/VegetableBiryani"; 
-import ButterChicken from "./pages/ButterChicken";
-import Prawn from "./pages/Prawn";
-import Fish from "./pages/Fish";
-import Biryani from "./pages/Biryani";
-import Thandai from "./pages/Thandai";
-import GulabJamun from "./pages/GulabJamun";
-import Rasgulla from "./pages/Rasgulla";
-import Jalebi from "./pages/Jalebi";
-import Falooda from "./pages/Falooda";
-import Lassi from "./pages/Lassi";
-import Coffee from "./pages/Coffee";
-import MasalaChai from "./pages/MasalaChai";
+import PaneerButterMasala from "./components/recipes/PaneerButterMasala";
+import DalMaKhani from "./components/recipes/DalMakhani";
+import PalakSoup from "./components/recipes/PalakSoup";
+import VegetableBiryani from "./components/recipes/VegetableBiryani"; 
+import ButterChicken from "./components/recipes/ButterChicken";
+import Prawn from "./components/recipes/Prawn";
+import Fish from "./components/recipes/Fish";
+import Biryani from "./components/recipes/Biryani";
+import Thandai from "./components/recipes/Thandai";
+import GulabJamun from "./components/recipes/GulabJamun";
+import Rasgulla from "./components/recipes/Rasgulla";
+import Jalebi from "./components/recipes/Jalebi";
+import Falooda from "./components/recipes/Falooda";
+import Lassi from "./components/recipes/Lassi";
+import Coffee from "./components/recipes/Coffee";
+import MasalaChai from "./components/recipes/MasalaChai";
 import UserProfile from "./pages/UserProfile";
 import AddRecipe from "./pages/AddRecipe";
 import About from "./pages/About";
