+++ frontend/src/components/recipes/Falooda.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { FaArrowLeft, FaHeart, FaRegHeart } from "react-icons/fa";
-import "../styles/RecipeDetail.css";
+import "../../styles/RecipeDetail.css";
 
-function Jalebi() {
+function Falooda() {
   const navigate = useNavigate();
   const [liked, setLiked] = useState(false);
   const [comments, setComments] = useState([]);
@@ -33,65 +33,48 @@ function Jalebi() {
         <FaArrowLeft /> Back
       </div>
 
-      <h1 className="text-2xl block text-center my-10">Jalebi</h1>
+      <h1 className="text-2xl block text-center my-10">Falooda</h1>
 
       <div className="recipe-image">
-        <img src="/jalebi.jpg" alt="Jalebi" />
+        <img src="/falooda.jpg" alt="Falooda" />
       </div>
 
       <div className="recipe-info">
         <h2 className="ingredientsH2 mt-2">About</h2>
         <p>
-          A crispy, deep-fried Indian dessert soaked in sugar syrup, known for
-          its spiral shape and sweet, tangy flavor.
+          A rich and refreshing Indian dessert drink made with milk, basil
+          seeds, falooda sev, rose syrup, and ice cream.
         </p>
 
-        <h2 className="ingredientsH2 mt-2 mb-2">Ingredients</h2>
-        <h3 className="ingredientsH2 mt-2">For the Jalebi Batter:</h3>
+        <h2 className="ingredientsH2 mt-2">Ingredients</h2>
         <ul className="ingredientsUl">
-          <li>1 cup All-purpose flour (maida)</li>
-          <li>2 tbsp Corn flour</li>
-          <li>1/2 cup Yogurt</li>
-          <li>1/2 tsp Baking soda</li>
-          <li>1/2 cup Warm water (as needed)</li>
-          <li>1/2 tsp Turmeric or saffron for color (optional)</li>
-          <li>Oil or ghee for deep frying</li>
-        </ul>
-
-        <h3 className="ingredientsH2 mt-2">For the Sugar Syrup:</h3>
-        <ul className="ingredientsUl">
-          <li>1 cup Sugar</li>
-          <li>1/2 cup Water</li>
-          <li>1/2 tsp Cardamom powder</li>
-          <li>1/2 tsp Lemon juice</li>
-          <li>1/2 tsp Rose water (optional)</li>
+          <li>2 cups Chilled milk</li>
+          <li>2 tbsp Rose syrup</li>
+          <li>2 tbsp Basil seeds (sabja), soaked</li>
+          <li>1/4 cup Falooda sev (vermicelli), cooked</li>
+          <li>2 tbsp Sugar (adjust to taste)</li>
+          <li>1 scoop Vanilla or kulfi ice cream</li>
+          <li>1 tbsp Chopped nuts (almonds/pistachios) for garnish</li>
+          <li>1/2 tbsp Chopped jelly cubes (optional)</li>
+          <li>Chopped fruits (like mango or strawberry, optional)</li>
         </ul>
 
         <h2 className="ingredientsH2 mt-2">Preparation Steps</h2>
         <ol className="list-decimal">
+          <li>Soak basil seeds in water for 10-15 minutes until they expand</li>
           <li>
-            In a bowl, mix all-purpose flour, corn flour, yogurt, and baking
-            soda
+            Boil falooda sev as per package instructions, then drain and cool
           </li>
+          <li>In a glass, add rose syrup at the bottom</li>
           <li>
-            Gradually add warm water to make a smooth, thick batter and let it
-            ferment for 6-8 hours
+            Layer soaked basil seeds, cooked falooda sev, and chilled milk
           </li>
+          <li>Add sugar if needed and mix gently</li>
           <li>
-            For sugar syrup: Boil sugar and water until sticky (1-string
-            consistency), then add lemon juice, cardamom powder, and rose water
+            Top with a scoop of ice cream, chopped nuts, and jelly cubes (if
+            using)
           </li>
-          <li>
-            Heat oil or ghee in a pan, fill a piping bag or squeeze bottle with
-            the batter, and pipe spirals into hot oil
-          </li>
-          <li>
-            Fry until golden and crisp, then dip immediately into warm sugar
-            syrup for 30 seconds
-          </li>
-          <li>
-            Remove and serve hot, garnished with saffron strands or chopped nuts
-          </li>
+          <li>Serve immediately with a long spoon and straw</li>
         </ol>
       </div>
 
@@ -132,4 +115,4 @@ function Jalebi() {
   );
 }
 
-export default Jalebi;
+export default Falooda;
