+++ frontend/src/components/recipes/Fish.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { FaArrowLeft, FaHeart, FaRegHeart } from "react-icons/fa";
-import "../styles/RecipeDetail.css";
+import "../../styles/RecipeDetail.css";
 
-function Prawn() {
+function Fish() {
   const navigate = useNavigate();
   const [liked, setLiked] = useState(false);
   const [comments, setComments] = useState([]);
@@ -27,57 +27,65 @@ function Prawn() {
   return (
     <div className="recipe-detail-container ingrdientsDiv">
       <div
-        className="back-button dark:text-white "
+        className="back-button dark:text-white"
         onClick={() => navigate("/nonveg")}
       >
         <FaArrowLeft /> Back
       </div>
 
-      <h1 className="text-2xl block text-center my-10">Prawn Masala</h1>
+      <h1 className="text-2xl block text-center my-10">Fish Curry</h1>
 
       <div className="recipe-image">
-        <img src="/prawn.jpg" alt="Prawn Masala" />
+        <img src="/fish.jpg" alt="Fish Curry" />
       </div>
 
       <div className="recipe-info">
         <h2 className="ingredientsH2">About</h2>
         <p>
-          A spicy and flavorful Indian dish made with prawns cooked in a rich
-          tomato-based masala with aromatic spices.
+          A delicious and aromatic Indian fish curry made with tender fish
+          pieces simmered in a spiced coconut or tomato-based gravy.
         </p>
 
         <h2 className="ingredientsH2">Ingredients</h2>
         <ul className="ingredientsUl">
-          <li>500g Prawns, cleaned and deveined</li>
+          <li>
+            500g Fish (Pomfret, Kingfish, or any firm fish), cut into pieces
+          </li>
           <li>2 tbsp Oil</li>
           <li>1 Onion, finely chopped</li>
           <li>2 Tomatoes, pureed</li>
           <li>1 tbsp Ginger-garlic paste</li>
+          <li>1 cup Coconut milk (or water for a lighter version)</li>
           <li>1 tsp Turmeric powder</li>
           <li>1 tsp Red chili powder</li>
-          <li>1 tsp Garam masala</li>
+          <li>1 tsp Coriander powder</li>
           <li>1/2 tsp Cumin powder</li>
-          <li>1/2 tsp Coriander powder</li>
+          <li>1 tsp Garam masala</li>
           <li>Salt to taste</li>
-          <li>1 tsp Kasuri methi (dried fenugreek leaves)</li>
+          <li>1 tsp Mustard seeds</li>
+          <li>1 Sprig Curry leaves</li>
+          <li>1 tsp Tamarind paste (optional, for tanginess)</li>
           <li>Fresh coriander for garnish</li>
         </ul>
 
         <h2 className="ingredientsH2">Preparation Steps</h2>
         <ol className="list-decimal">
-          <li>Heat oil in a pan and sauté onions until golden brown</li>
-          <li>Add ginger-garlic paste and cook until fragrant</li>
+          <li>
+            Heat oil in a pan, add mustard seeds and curry leaves, and let them
+            splutter
+          </li>
+          <li>Sauté onions until golden brown, then add ginger-garlic paste</li>
           <li>
             Mix in tomato puree, turmeric, chili powder, cumin, and coriander
             powder
           </li>
           <li>Cook until the oil separates from the masala</li>
+          <li>Add coconut milk and tamarind paste, mix well</li>
           <li>
-            Add prawns, salt, and cook for 5-7 minutes until they turn pink and
-            tender
+            Gently add fish pieces, cover, and cook for 10 minutes until tender
           </li>
-          <li>Stir in garam masala and kasuri methi, mix well</li>
-          <li>Garnish with fresh coriander and serve hot with rice or roti</li>
+          <li>Sprinkle garam masala and garnish with fresh coriander</li>
+          <li>Serve hot with steamed rice or roti</li>
         </ol>
       </div>
 
@@ -93,7 +101,7 @@ function Prawn() {
         </div>
 
         <div className="comment-section">
-          <h3 className="ingredientsH">Comments</h3>
+          <h3>Comments</h3>
           <form onSubmit={handleComment}>
             <textarea
               value={newComment}
@@ -118,4 +126,4 @@ function Prawn() {
   );
 }
 
-export default Prawn;
+export default Fish;
