+++ frontend/src/components/recipes/PalakSoup.jsx
@@ -94,7 +94,7 @@ function PalakSoup() {
           />
           <button
             type="submit"
-            className="mt-3 bg-red-500 hover:bg-red-600 text-white px-6 py-2 rounded-full text-black"
+            className="mt-3 bg-red-500 hover:bg-red-600 text-white px-6 py-2 rounded-full"
           >
             Post Comment
           </button>
