+++ frontend/src/components/recipes/Prawn.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { FaArrowLeft, FaHeart, FaRegHeart } from "react-icons/fa";
-import "../styles/RecipeDetail.css";
+import "../../styles/RecipeDetail.css";
 
-function Falooda() {
+function Prawn() {
   const navigate = useNavigate();
   const [liked, setLiked] = useState(false);
   const [comments, setComments] = useState([]);
@@ -27,54 +27,57 @@ function Falooda() {
   return (
     <div className="recipe-detail-container ingrdientsDiv">
       <div
-        className="back-button !text-black dark:!text-white"
-        onClick={() => navigate("/dessert")}
+        className="back-button dark:text-white "
+        onClick={() => navigate("/nonveg")}
       >
         <FaArrowLeft /> Back
       </div>
 
-      <h1 className="text-2xl block text-center my-10">Falooda</h1>
+      <h1 className="text-2xl block text-center my-10">Prawn Masala</h1>
 
       <div className="recipe-image">
-        <img src="/falooda.jpg" alt="Falooda" />
+        <img src="/prawn.jpg" alt="Prawn Masala" />
       </div>
 
       <div className="recipe-info">
-        <h2 className="ingredientsH2 mt-2">About</h2>
+        <h2 className="ingredientsH2">About</h2>
         <p>
-          A rich and refreshing Indian dessert drink made with milk, basil
-          seeds, falooda sev, rose syrup, and ice cream.
+          A spicy and flavorful Indian dish made with prawns cooked in a rich
+          tomato-based masala with aromatic spices.
         </p>
 
-        <h2 className="ingredientsH2 mt-2">Ingredients</h2>
+        <h2 className="ingredientsH2">Ingredients</h2>
         <ul className="ingredientsUl">
-          <li>2 cups Chilled milk</li>
-          <li>2 tbsp Rose syrup</li>
-          <li>2 tbsp Basil seeds (sabja), soaked</li>
-          <li>1/4 cup Falooda sev (vermicelli), cooked</li>
-          <li>2 tbsp Sugar (adjust to taste)</li>
-          <li>1 scoop Vanilla or kulfi ice cream</li>
-          <li>1 tbsp Chopped nuts (almonds/pistachios) for garnish</li>
-          <li>1/2 tbsp Chopped jelly cubes (optional)</li>
-          <li>Chopped fruits (like mango or strawberry, optional)</li>
+          <li>500g Prawns, cleaned and deveined</li>
+          <li>2 tbsp Oil</li>
+          <li>1 Onion, finely chopped</li>
+          <li>2 Tomatoes, pureed</li>
+          <li>1 tbsp Ginger-garlic paste</li>
+          <li>1 tsp Turmeric powder</li>
+          <li>1 tsp Red chili powder</li>
+          <li>1 tsp Garam masala</li>
+          <li>1/2 tsp Cumin powder</li>
+          <li>1/2 tsp Coriander powder</li>
+          <li>Salt to taste</li>
+          <li>1 tsp Kasuri methi (dried fenugreek leaves)</li>
+          <li>Fresh coriander for garnish</li>
         </ul>
 
-        <h2 className="ingredientsH2 mt-2">Preparation Steps</h2>
+        <h2 className="ingredientsH2">Preparation Steps</h2>
         <ol className="list-decimal">
-          <li>Soak basil seeds in water for 10-15 minutes until they expand</li>
+          <li>Heat oil in a pan and sauté onions until golden brown</li>
+          <li>Add ginger-garlic paste and cook until fragrant</li>
           <li>
-            Boil falooda sev as per package instructions, then drain and cool
+            Mix in tomato puree, turmeric, chili powder, cumin, and coriander
+            powder
           </li>
-          <li>In a glass, add rose syrup at the bottom</li>
+          <li>Cook until the oil separates from the masala</li>
           <li>
-            Layer soaked basil seeds, cooked falooda sev, and chilled milk
+            Add prawns, salt, and cook for 5-7 minutes until they turn pink and
+            tender
           </li>
-          <li>Add sugar if needed and mix gently</li>
-          <li>
-            Top with a scoop of ice cream, chopped nuts, and jelly cubes (if
-            using)
-          </li>
-          <li>Serve immediately with a long spoon and straw</li>
+          <li>Stir in garam masala and kasuri methi, mix well</li>
+          <li>Garnish with fresh coriander and serve hot with rice or roti</li>
         </ol>
       </div>
 
@@ -82,7 +85,7 @@ function Falooda() {
         <div className="mt-10 flex items-center gap-4">
           <button
             onClick={handleLike}
-            className="flex items-center gap-2 text-gray-600 dark:text-white hover:text-red-500 dark:hover:text-red-500 focus:outline-none "
+            className="flex items-center gap-2 text-gray-600 dark:text-white hover:text-red-500 dark:hover:text-red-500 focus:outline-none"
           >
             {liked ? <FaHeart className="text-red-500" /> : <FaRegHeart />}{" "}
             {liked ? "Liked" : "Like"}
@@ -90,7 +93,7 @@ function Falooda() {
         </div>
 
         <div className="comment-section">
-          <h3>Comments</h3>
+          <h3 className="ingredientsH">Comments</h3>
           <form onSubmit={handleComment}>
             <textarea
               value={newComment}
@@ -115,4 +118,4 @@ function Falooda() {
   );
 }
 
-export default Falooda;
+export default Prawn;
