+++ frontend/src/components/recipes/Rasgulla.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { FaArrowLeft, FaHeart, FaRegHeart } from 'react-icons/fa';
-import '../styles/RecipeDetail.css';
+import '../../styles/RecipeDetail.css';
 
-function Coffee() {
+function Rasgulla() {
   const navigate = useNavigate();
   const [liked, setLiked] = useState(false);
   const [comments, setComments] = useState([]);
@@ -23,36 +23,48 @@ function Coffee() {
 
   return (
     <div className="recipe-detail-container ingrdientsDiv">
-      <div className="back-button  !text-black dark:!text-white" onClick={() => navigate('/beverages')}>
+      <div className="back-button !text-black dark:!text-white" onClick={() => navigate('/dessert')}>
         <FaArrowLeft /> Back
       </div>
       
-      <h1 className="text-2xl block text-center my-10">Filter Coffee</h1>
+      <h1  className="text-2xl block text-center my-10">Rasgulla</h1>
       
       <div className="recipe-image">
-        <img src="/coffee.jpg" alt="Filter Coffee" />
+        <img src="/rasgulla.jpg" alt="Rassgulla" />
       </div>
 
       <div className="recipe-info">
-        <h2 className="ingredientsH2 mt-2">About</h2>
-        <p>A rich and aromatic South Indian coffee made using a traditional filter, blended with frothy milk for a smooth and bold taste.</p>
+        <h2 className='ingredientsH2 mt-2'>About</h2>
+        <p>A soft and spongy Bengali dessert made from chhena (Indian cottage cheese) balls cooked in a light sugar syrup.</p>
 
-        <h2 className="ingredientsH2 mt-2">Ingredients</h2>
-        <ul className="ingredientsUl">
-          <li>2 tbsp Ground coffee powder (South Indian filter coffee blend)</li>
-          <li>1 cup Water</li>
-          <li>1/2 cup Boiled milk</li>
-          <li>1-2 tsp Sugar (adjust to taste)</li>
+        <h2 className='ingredientsH2 mt-2 mb-2'>Ingredients : </h2>
+        <h3 className='ingredientsH2 mt-2'>For the Chhena (Paneer):</h3>
+        <ul className='ingredientsUl'>
+          <li>1 liter Full-fat milk</li>
+          <li>2 tbsp Lemon juice or vinegar</li>
+          <li>2 cups Cold water</li>
+          <li>1 tsp Corn flour (optional, for binding)</li>
         </ul>
 
-        <h2 className="ingredientsH2 mt-2">Preparation Steps</h2>
-        <ol class="list-decimal">
-          <li>Add coffee powder to the upper chamber of a traditional South Indian coffee filter</li>
-          <li>Pour hot water over the coffee and let it drip into the lower chamber (takes about 10-15 minutes)</li>
-          <li>Boil milk and mix with the brewed coffee decoction</li>
-          <li>Add sugar as per taste and mix well</li>
-          <li>Pour the coffee back and forth between two cups to create froth</li>
-          <li>Serve hot in a traditional tumbler and dabarah (small bowl) for authentic taste</li>
+        <h3 className='ingredientsH2 mt-2'>For the Sugar Syrup:</h3>
+        <ul className='ingredientsUl'>
+          <li>1 cup Sugar</li>
+          <li>4 cups Water</li>
+          <li>2 Cardamom pods (optional)</li>
+          <li>1 tsp Rose water (optional)</li>
+        </ul>
+
+        <h2 className='ingredientsH2 mt-2'>Preparation Steps</h2>
+        <ol className='list-decimal'>
+          <li>Boil milk, then add lemon juice or vinegar to curdle it</li>
+          <li>Strain the curdled milk using a muslin cloth and rinse with cold water to remove acidity</li>
+          <li>Hang the chhena for 30 minutes to remove excess water</li>
+          <li>Knead the chhena for 5-7 minutes until smooth and soft</li>
+          <li>Add corn flour (if using) and knead again, then shape into small, smooth balls</li>
+          <li>In a wide pan, boil water and sugar to make syrup, adding cardamom pods</li>
+          <li>Drop the chhena balls into boiling syrup and cook covered for 15-20 minutes on medium heat</li>
+          <li>Turn off the heat and let them cool in the syrup</li>
+          <li>Add rose water if desired and refrigerate before serving chilled</li>
         </ol>
       </div>
 
@@ -95,4 +107,4 @@ function Coffee() {
   );
 }
 
-export default Coffee;
+export default Rasgulla;
