+++ frontend/src/components/recipes/Thandai.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { FaArrowLeft, FaHeart, FaRegHeart } from "react-icons/fa";
-import "../styles/RecipeDetail.css";
+import "../../styles/RecipeDetail.css";
 
-function Fish() {
+function Thandai() {
   const navigate = useNavigate();
   const [liked, setLiked] = useState(false);
   const [comments, setComments] = useState([]);
@@ -27,65 +27,61 @@ function Fish() {
   return (
     <div className="recipe-detail-container ingrdientsDiv">
       <div
-        className="back-button dark:text-white"
-        onClick={() => navigate("/nonveg")}
+        className="back-button !text-black dark:!text-white"
+        onClick={() => navigate("/beverages")}
       >
         <FaArrowLeft /> Back
       </div>
 
-      <h1 className="text-2xl block text-center my-10">Fish Curry</h1>
+      <h1 className="text-2xl block text-center my-10">Thandai</h1>
 
       <div className="recipe-image">
-        <img src="/fish.jpg" alt="Fish Curry" />
+        <img src="/thandai.jpg" alt="Thandai" />
       </div>
 
       <div className="recipe-info">
-        <h2 className="ingredientsH2">About</h2>
+        <h2 className="ingredientsH2 mt-2">About</h2>
         <p>
-          A delicious and aromatic Indian fish curry made with tender fish
-          pieces simmered in a spiced coconut or tomato-based gravy.
+          A refreshing and aromatic Indian beverage made with milk, nuts, and a
+          blend of spices, traditionally enjoyed during festivals like Holi and
+          Mahashivratri.
         </p>
 
-        <h2 className="ingredientsH2">Ingredients</h2>
+        <h2 className="ingredientsH2 mt-2">Ingredients</h2>
         <ul className="ingredientsUl">
-          <li>
-            500g Fish (Pomfret, Kingfish, or any firm fish), cut into pieces
-          </li>
-          <li>2 tbsp Oil</li>
-          <li>1 Onion, finely chopped</li>
-          <li>2 Tomatoes, pureed</li>
-          <li>1 tbsp Ginger-garlic paste</li>
-          <li>1 cup Coconut milk (or water for a lighter version)</li>
-          <li>1 tsp Turmeric powder</li>
-          <li>1 tsp Red chili powder</li>
-          <li>1 tsp Coriander powder</li>
-          <li>1/2 tsp Cumin powder</li>
-          <li>1 tsp Garam masala</li>
-          <li>Salt to taste</li>
-          <li>1 tsp Mustard seeds</li>
-          <li>1 Sprig Curry leaves</li>
-          <li>1 tsp Tamarind paste (optional, for tanginess)</li>
-          <li>Fresh coriander for garnish</li>
+          <li>2 cups Milk</li>
+          <li>2 tbsp Almonds</li>
+          <li>1 tbsp Cashews</li>
+          <li>1 tbsp Pistachios</li>
+          <li>1 tbsp Poppy seeds</li>
+          <li>1 tbsp Fennel seeds</li>
+          <li>1/2 tsp Black pepper</li>
+          <li>4-5 Cardamom pods</li>
+          <li>2 tbsp Sugar or honey</li>
+          <li>1/2 tsp Rose water (optional)</li>
+          <li>1/2 tsp Saffron strands (optional)</li>
+          <li>Chilled water or ice cubes (optional, for a lighter version)</li>
+          <li>Chopped nuts for garnish</li>
         </ul>
 
-        <h2 className="ingredientsH2">Preparation Steps</h2>
+        <h2 className="ingredientsH2 mt-2">Preparation Steps</h2>
         <ol className="list-decimal">
           <li>
-            Heat oil in a pan, add mustard seeds and curry leaves, and let them
-            splutter
+            Soak almonds, cashews, pistachios, poppy seeds, fennel seeds, and
+            cardamom in warm water for 2-3 hours
           </li>
-          <li>Sauté onions until golden brown, then add ginger-garlic paste</li>
           <li>
-            Mix in tomato puree, turmeric, chili powder, cumin, and coriander
-            powder
+            Blend the soaked ingredients with black pepper and a little milk to
+            form a smooth paste
           </li>
-          <li>Cook until the oil separates from the masala</li>
-          <li>Add coconut milk and tamarind paste, mix well</li>
+          <li>Boil milk, add the prepared paste, and mix well</li>
           <li>
-            Gently add fish pieces, cover, and cook for 10 minutes until tender
+            Stir in sugar, saffron, and rose water (if using), and let it cool
           </li>
-          <li>Sprinkle garam masala and garnish with fresh coriander</li>
-          <li>Serve hot with steamed rice or roti</li>
+          <li>Refrigerate for 1-2 hours for flavors to blend</li>
+          <li>
+            Strain (optional) and serve chilled, garnished with chopped nuts
+          </li>
         </ol>
       </div>
 
@@ -93,7 +89,7 @@ function Fish() {
         <div className="mt-10 flex items-center gap-4">
           <button
             onClick={handleLike}
-            className="flex items-center gap-2 text-gray-600 dark:text-white hover:text-red-500 dark:hover:text-red-500 focus:outline-none"
+            className="flex items-center gap-2 text-gray-600 dark:text-white hover:text-red-500 dark:hover:text-red-500 focus:outline-none "
           >
             {liked ? <FaHeart className="text-red-500" /> : <FaRegHeart />}{" "}
             {liked ? "Liked" : "Like"}
@@ -126,4 +122,4 @@ function Fish() {
   );
 }
 
-export default Fish;
+export default Thandai;
