+++ frontend/src/components/recipes/VegetableBiryani.jsx
