+++ frontend/src/pages/About.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './About.css';
+import '../styles/About.css'; // Assuming you have a CSS file for styling
 
 const About = () => {
   return (
