+++ frontend/src/styles/Error.css
